pyproject.toml
[tool.poetry]
name = "django-cowsay"
version = "0.1.0"
description = "What does the cow say?"
authors = ["Joseph Brown <joseph.d.brown94@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
django-simple-history = "^2.8.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
